style refactor
